refactor common CustomRagingAura off Path of the Elements and Path of the Spirits
